Refactor export CSV / TSV download menu to a new file
